Sequencer UI layout tweaks.
There was a mix in the layout of these panels; some were single column, some dual column. All these n-key sidebars seem to work best with single column layouts to minimize wasted space, so made all the panels fit that paradigm.

Also included a bunch of labels and a bit of layout cleanup.