Added 'Evaluation Time' setting access to Path Animation panel.
This was previously only available from the datablocks viewer, but this is the direct replacement to the badly named 'speed' ipo-curve in the past ('speed' implies a rate that must be integrated/added to the results of past frame, rather than a factor).

Also, tweaked the RNA definition so that this shows as a slider (i.e. a factor), since the valid values for this are clamped to the [0, 1].